fix: extend nilla projects, use path as src
The Nilla loader was incorrectly processing things like module overrides
which should have been passed along to the extend function. This commit
corrects this behavior with the caveat that the type is now perhaps not
specific enough. It may be necessary in the future to specify the full
structure of `extend` attributes to allow for some intelligent merging
of the base level while still allowing unevaluated module system
integrations to be passed through.

The input path type was also overly restrictive, resulting in an
inability to provide inputs via a path. This was unnecessary and broke
several reasonable use cases. The type has been updated to now allow
either derivations or paths.

FIXES #14
FIXES #17