fix: cut eval early for __functor
__functor does not have a src and because it doesn't have a loader set
nilla tries to determine which one to use. That action depends on the src
attribute which causes all evals to fail. This patch fixes this by
setting the loader to not exist if the input attribute name is __functor,
this may be edited later to be more general (that is, for anything which
doesn't have a src attribute).